# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "plot api",
    "contact": {}
  },
  "host": "localhost:4000",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/skill": {
      "post": {
        "summary": "Create Skill",
        "tags": ["Skills"],
        "operationId": "CreateSkill",
        "deprecated": false,
        "produces": ["application/json; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateSkillRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateSkill"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "status": "Success",
                "message": "Create Skill Talking Successfully"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "66"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"42-Q0/EBBhcdiE/DBlEOK2pFWYOH5k\""
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              },
              "Date": {
                "type": "string",
                "default": "Wed, 29 Sep 2021 17:06:16 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Skill",
        "tags": ["Skills"],
        "operationId": "DeleteSkill",
        "deprecated": false,
        "produces": ["application/json; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/DeleteSkillRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DeleteSkill"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Deleted Skill With Name Talking"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "45"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"2d-QFklHi0X3nhMe5k2rWvOhre6XGA\""
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              },
              "Date": {
                "type": "string",
                "default": "Wed, 29 Sep 2021 17:06:59 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
...
    "/subject/5f8c7ee1b9776e05f105a6db": {
      "get": {
        "summary": "Subject",
        "tags": ["Subject"],
        "operationId": "Subject",
        "deprecated": false,
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/sms/allstudent": {
      "get": {
        "summary": "get all student phone number",
        "tags": ["Sms"],
        "operationId": "getallstudentphonenumber",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/sms/send": {
      "post": {
        "summary": "Send Sms",
        "tags": ["Sms"],
        "operationId": "SendSms",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SendSmsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateSkillRequest": {
      "title": "CreateSkillRequest",
      "example": {
        "name": "Talking",
        "school": "5f9183cad035174e94992476"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "school": {
          "type": "string"
        }
      },
      "required": ["name", "school"]
    },
    "CreateSkill": {
      "title": "CreateSkill",
      "example": {
        "status": "Success",
        "message": "Create Skill Talking Successfully"
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      },
      "required": ["status", "message"]
    },
...
    "DeleteSkillRequest": {
      "title": "DeleteSkillRequest",
      "example": {
        "skillId": "61549d08e846868796bc6c6f",
        "school": "5f9183cad035174e94992476"
      },
      "type": "object",
      "properties": {
        "skillId": {
          "type": "string"
        },
        "school": {
          "type": "string"
        }
      },
      "required": ["skillId", "school"]
    },
    "DeleteSkill": {
      "title": "DeleteSkill",
      "example": {
        "message": "Deleted Skill With Name Talking"
      },
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": ["message"]
    },
...
    "SendSmsRequest": {
      "title": "SendSmsRequest",
      "example": {
        "phoneNumbers": ["08144899798"],
        "message": "Hi it the us gggggggg  and we would like to do that we do best make the xhshshsa hsycs . hcscwb . shsshvcsn . zcshcbsns sncbsc sscbsc s cshcbscs",
        "school": "5f9183cad035174e94992476"
      },
      "type": "object",
      "properties": {
        "phoneNumbers": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "message": {
          "type": "string"
        },
        "school": {
          "type": "string"
        }
      },
      "required": ["phoneNumbers", "message", "school"]
    },
...
  },
...
}